Pull Request: Major Modifications in main.py
Description:
This pull request introduces significant modifications to the main.py file, enhancing its functionality and improving user experience through the implementation of the extraction class and the addition of a directory selection button.

Changes:

**Utilization of Extraction Class:**
Replaced previous code and using extraction.py class in main.py. This modification improves code readability, maintainability, and allows for easier expansion of extraction functionalities in the future.

**Addition of Directory Selection Button:**
Implemented a new button within the user interface to offer users the choice to select a directory instead of just a file. If the user chose the directory option it will get automatically the title of the sheet music and name the pdf file.